requirements.txt
PyYAML==5.3.1
requests==2.22.0 
numpy==1.24.2
scipy==1.8.0
yahist==1.14.0
mplhep==0.3.26
tqdm==4.64.1
gitpython==3.1.29
tabulate==0.9.0
aiofiles==22.1.0
rich==13.3.0
pandas==1.5.2
awkward==2.0.7
pyarrow==11.0.0
emoji==2.2.0
flask==2.2.5
hist==2.6.3
git+https://github.com/danbarto/cocina.git@v1.0.0
colored==1.4.4
